agendaUsuario terminado, arreglo de campos tipo time y tipo date.

+++ src/main/java/sena/petcom/model/Usuario/Usuario.java
@@ -54,7 +54,7 @@ public class Usuario {
 
     @NotNull
     @Column(name="telefonoUsu", unique=true)
-    private Integer telefonoUsu;
+    private Long telefonoUsu;
 
     @NotEmpty
     @NotNull
